Feat: Add POST /users
Add endpoint to create a user with the following data: FirstName(string), lastName(string), phone(string), password (string), tosAgreement (bool)

The idea is that a user can only have a distinct phone number, no two users will have the same phone.
The password will be SHA256 hashed, so a new helpers file was created to handle this, along with the parsing of a json to object.